# perf: add Vector::Add, Sub and ScalarMul assembly (and purego) implementations