--- a/ecc/bls12-377/fp/vector.go
+++ b/ecc/bls12-377/fp/vector.go
@@ -213,6 +213,12 @@ func (vector *Vector) Sub(a, b Vector) {
 	vector.subGeneric(a, b)
 }
 
+// ScalarMul multiplies a vector by a scalar element-wise and stores the result in self.
+// It panics if the vectors don't have the same length.
+func (vector *Vector) ScalarMul(a Vector, b *Element) {
+	vector.scalarMulGeneric(a, b)
+}
+
 func (vector *Vector) addGeneric(a, b Vector) {
 	if len(a) != len(b) || len(a) != len(*vector) {
 		panic("vector.Add: vectors don't have the same length")
@@ -228,6 +234,15 @@ func (vector *Vector) subGeneric(a, b Vector) {
 	}
 	for i := 0; i < len(a); i++ {
 		(*vector)[i].Sub(&a[i], &b[i])
+	}
+}
+
+func (vector *Vector) scalarMulGeneric(a Vector, b *Element) {
+	if len(a) != len(*vector) {
+		panic("vector.ScalarMul: vectors don't have the same length")
+	}
+	for i := 0; i < len(a); i++ {
+		(*vector)[i].Mul(&a[i], b)
 	}
 }