--- a/ecc/bls12-377/fr/element_ops_amd64.go
+++ b/ecc/bls12-377/fr/element_ops_amd64.go
@@ -45,6 +45,11 @@ func reduce(res *Element)
 //go:noescape
 func Butterfly(a, b *Element)
 
+// AddVec res = a + b
+//
+//go:noescape
+func AddVec(res, a, b *Element, n uint64)
+
 // Mul z = x * y (mod q)
 //
 // x and y must be less than q