--- a/ecc/bls12-377/fr/element_ops_amd64.go
+++ b/ecc/bls12-377/fr/element_ops_amd64.go
@@ -69,6 +69,18 @@ func (vector *Vector) Sub(a, b Vector) {
 //go:noescape
 func subVec(res, a, b *Element, n uint64)
 
+// ScalarMul multiplies a vector by a scalar element-wise and stores the result in self.
+// It panics if the vectors don't have the same length.
+func (vector *Vector) ScalarMul(a Vector, b *Element) {
+	if len(a) != len(*vector) {
+		panic("vector.ScalarMul: vectors don't have the same length")
+	}
+	scalarMulVec(&(*vector)[0], &a[0], b, uint64(len(a)))
+}
+
+//go:noescape
+func scalarMulVec(res, a, b *Element, n uint64)
+
 // Mul z = x * y (mod q)
 //
 // x and y must be less than q