--- a/ecc/bls12-377/fr/element_ops_purego.go
+++ b/ecc/bls12-377/fr/element_ops_purego.go
@@ -60,6 +60,12 @@ func reduce(z *Element) {
 	_reduceGeneric(z)
 }
 
+// Add adds two vectors element-wise and stores the result in self.
+// It panics if the vectors don't have the same length.
+func (vector *Vector) Add(a, b Vector) {
+	vector.addGeneric(a, b)
+}
+
 // Mul z = x * y (mod q)
 //
 // x and y must be less than q