--- a/ecc/bls12-377/fr/element_ops_purego.go
+++ b/ecc/bls12-377/fr/element_ops_purego.go
@@ -63,19 +63,19 @@ func reduce(z *Element) {
 // Add adds two vectors element-wise and stores the result in self.
 // It panics if the vectors don't have the same length.
 func (vector *Vector) Add(a, b Vector) {
-	vector.addGeneric(a, b)
+	addVecGeneric(*vector, a, b)
 }
 
 // Sub subtracts two vectors element-wise and stores the result in self.
 // It panics if the vectors don't have the same length.
 func (vector *Vector) Sub(a, b Vector) {
-	vector.subGeneric(a, b)
+	subVecGeneric(*vector, a, b)
 }
 
 // ScalarMul multiplies a vector by a scalar element-wise and stores the result in self.
 // It panics if the vectors don't have the same length.
 func (vector *Vector) ScalarMul(a Vector, b *Element) {
-	vector.scalarMulGeneric(a, b)
+	scalarMulVecGeneric(*vector, a, b)
 }
 
 // Mul z = x * y (mod q)