# Description

Adds 
```
// Add adds two vectors element-wise and stores the result in self.
// It panics if the vectors don't have the same length.
func (vector *Vector) Add(a, b Vector)

// Sub subtracts two vectors element-wise and stores the result in self.
// It panics if the vectors don't have the same length.
func (vector *Vector) Sub(a, b Vector)

// ScalarMul multiplies a vector by a scalar element-wise and stores the result in self.
// It panics if the vectors don't have the same length.
func (vector *Vector) ScalarMul(a Vector, b *Element)
```

Assembly (amd64 target only) is generated for modulus < 256 bits.

Benchmarks on `r7i.xlarge` of x86 assembly vs pure go path: (perf gain are less interesting on on some AMD chips like `hpc6a` instances)
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkElementVecOps/Add-4           2779          1762          -36.60%
BenchmarkElementVecOps/Add-4           2787          1768          -36.56%
BenchmarkElementVecOps/Sub-4           2767          1706          -38.34%
BenchmarkElementVecOps/Sub-4           2779          1715          -38.29%
BenchmarkElementVecOps/ScalarMul-4     18014         13515         -24.98%
BenchmarkElementVecOps/ScalarMul-4     18030         13533         -24.94%
```

Good starting point against something using AVX like [this](https://github.com/sragss/vectorized-fields).